Fix singularity in tetrahedron evaluation. Disabled OpenMP in micromag

+++ source/MagTenseMicroMag/MicroMagParameters.f90
@@ -105,7 +105,7 @@ include "mkl_dfti.f90"
         integer :: demag_approximation                          !> Flag for how to approximate the demagnetization tensor as specified in the parameters below
         integer :: demagTensorReturnState                       !> Flag describing how or if the demag tensor should be returned
         integer :: demagTensorLoadState                         !> Flag describing how or if to load the demag tensor (from disk e.g.)
-        integer*4 :: nThreadsMatlab                             !> Number of threads to use in the OpenMP demag tensor allocation
+        integer :: nThreadsMatlab                             !> Number of threads to use in the OpenMP demag tensor allocation
         integer,dimension(3) :: N_ave                           !> Number of points to average the demag tensor in in the recieving tile, N_ave(1) = N_x etc
         character*256 :: demagTensorFileOut, demagTensorFileIn  !> Filename (including path) for output (input) of demag tensor if it is to be returned as a file (demagTensorReturnState >2 and the value is equal to the length of the file including path)
